Move server related scripts to a folder and remove unnecessary test fixture
