+++ scripts/server/serverTest.js
@@ -1,5 +1,5 @@
 import baseTest from '@playwright/test'
-import { startServer } from '../scripts/modules/server.mjs'
+import { startServer } from './server.mjs'
 
 const serverTest = baseTest.extend({
     port: [async ({}, use, workerInfo) => { // eslint-disable-line no-empty-pattern
@@ -13,10 +13,6 @@ const serverTest = baseTest.extend({
         await use(server)
         await server.close()
     }, { scope: 'worker', auto: true }],
-    page: async ({ baseUrl, page }, use) => {
-        await page.goto(baseUrl)
-        await use(page)
-    },
 })
 
 export default serverTest
